Modularized Backend Structure
