+++ backend/middleware/errorHandler.js
@@ -1,0 +1,4 @@
+export const errorHandler = (err, req, res, next) => {
+  console.error(err.stack);
+  res.status(500).json({ success: false, message: "Internal Server Error", errors: err.message });
+}
