+++ backend/middleware/validation.js
@@ -17,13 +17,12 @@ const registerValidationRules = [
 const validate = (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    return res.status(400).json({
-      success: false,
-      message: "Validation failed",
-      errors: errors.array(),
-    });
+    return res.status(400).json({ errors: errors.array() });
   }
   next();
 };
 
-module.exports = { registerValidationRules, validate};
+module.exports = {
+  registerValidationRules,
+  validate,
+};
